Initial working version which reduces the stuttering caused by Bracken
This changes the process for a Bracken to find a far navmesh node to run to so that it runs in a coroutine which checks one node per frame. This prevents the >100ms stutter that can occur if all paths are blocked.

The behavior when all paths are blocked by line of sight has been changed to avoid that the Bracken ends up moving and stopping repeatedly as well. Instead of just picking the furthest node, it will now find the furthest accessible node, which prevents the NavMeshAgent from rejecting the destination.