Add a patch to remove lag when snare fleas are looking for a starting position
Since snare fleas try to find a path that is midway through the list of available paths as sorted by distance, they would spend a significant amount of time checking and then discarding paths. Since the node number in the list is among the valid paths and not all paths, we can't just skip pathfinding, so instead we amortize it like the rest of the patches.

This will also fix lag spikes that will occur when the snare fleas are outside the navmesh, which happens most often when they fall off.