Implement a new system to do node scanning off the main thread
Pathfinding is implemented through the NavMeshQuery API to allow it to safely run off the main thread.

The AI nodes are sorted using insertion sort and then cached to reduce the amount of work that is necessary to re-sort it with each node-choosing task.

When an AI needs to choose a node to path to, it will call one of the methods in AsyncPathfinding to start a job with an ID number to differentiate each type of pathfinding task for that AI. When any previous job is running with a different ID, it will wait for that job to complete and then start a new one with the new ID. Then, any subsequent calls will find a job running with that ID and return its status without stopping it. Once the job has found a node, the caller can use that node to set a new destination.

Raycasts for checking line of sight are still done on the main thread, one path per frame. This normally still allows the pathfinding to complete within the 200ms window between DoAIInterval() calls, but if it is delayed, that is preferable to dropping the framerate considerably by running through a bunch of paths in a frame.