Add a patch to simplify tulip snakes warping to navmesh when far away
If a tulip snake dismounts the player while >15 units away from the navmesh, they will call ChooseClosestNodeToPosition. Thankfully, this doesn't go as far as calculating paths since the tulip snake is guaranteed to be off the navmesh, but since we already know that no path will be valid, the call is equivalent to just choosing the nearest AI node with no other criteria.

Patching this to simply find the closest node prevents a few allocations and a sorting operation.