Ensure that player pathing isn't reset without retrieving the result
Since we retrieved the player pathing status twice, once to check whether a path to one player is valid, and a second time to check whether all path jobs have completed, it was possible for the path job to complete between getting the single path status and preparing to reset the pathfinding operation. Instead, keep track of whether we have retrieved each job status after it has completed, then reset the operation only when all have been retrieved.

Additionally, the reset is delayed until the end of the frame to ensure that if TargetClosestPlayer is called twice in one frame, it will not always return false on the second call.