Correct the checkLOSToPosition behavior to match the vanilla use case
Only the cave dweller uses this parameter, and the reproduction of the code was based on the branch of the vanilla method that doesn't calculate the path distance. The branch that calculates the path distance, which the cave uses, differs from that code slightly, so let's make it totally consistent.